Add files via upload
When developers started experimenting with ways to write and exchange code online in the early days of the internet,
the history of online code editors began. Leo Horie, a software developer, designed the "JavaScript Sandbox" in 2004,
one of the first instances of an online code editor, allowing users to write and run JavaScript code directly in their web
browsers.
Online code editors have grown more powerful and complex over time, with numerous new features and functionalities
being added to enhance user experience.
One of the most important advancements in this area was the invention of collaborative online code editors, which let
several developers to operate on the same code in real-time.